refactor(server): extract connection handling into server_conn.go
Move serveListener, serveConn, processOneRequest, parseServerTimeout,
isShutdown, closeConn, readRequest, auth out of server.go into
server_conn.go. Same package, mechanical move.
